requirements.txt
schedule
tweepy
python-dotenv
text-to-tweets